backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [

]
readme = "README.md"
python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "requests (>=2.32.5,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
